# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a simple API definition for the Autoamated Library Management\
    \ System"
  version: "1.0.0"
  title: "Automated Library Management System"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
  /login:
    post:
      tags:
      - "user"
      summary: "Logs in and returns the authentication  cookie"
      operationId: "loginPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "loginRequest"
        description: "A JSON object containing the login and password."
        required: true
        schema:
          $ref: "#/definitions/LoginRequest"
      responses:
        200:
          description: "Successfully authenticated. The session ID is returned in\
            \ a cookie named `X-SESSION`. You need to include this cookie in subsequent\
            \ requests.\n"
          headers:
            Set-Cookie:
              type: "string"
              description: "Set the session token in the cookie"
              default: "X-SESSION=sessionotkenhere"
        400:
          description: "Bad Request\n"
      x-swagger-router-controller: "User"
...
securityDefinitions:
  SessionKeyAuth:
    type: "apiKey"
    name: "X-SESSION"
    in: "header"
definitions:
...
  LoginRequest:
    required:
    - "password"
    - "username"
    properties:
      username:
        type: "string"
        minLength: 1
      password:
        type: "string"
        minLength: 1
...